pyproject.toml
[tool.poetry]
name = "py-html-gen"
version = "0.1.0"
description = ""
authors = ["Logan Donley <logan.donley@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
